+++ src/libs/ajax.js
@@ -299,9 +299,14 @@ const Billing = signal => ({
     const res = await fetchSam('api/resources/v1/billing-project', _.merge(authOpts(), { signal }))
     return res.json()
   },
-  listUsersOnBillingProject: async projectId => {
-    const res = await fetchRawls(`billing/${projectId}/members`, _.merge(authOpts(), { signal }))
-    return res.json()
+
+  project: projectId => {
+    return {
+      listUsers: async () => {
+        const res = await fetchRawls(`billing/${projectId}/members`, _.merge(authOpts(), { signal }))
+        return res.json()
+      }
+    }
   }
 })
